Teams Switch Fully Working + BugFix
Teams Switch button is fully working now :) Including rejecting switch if the other team is full.

This had a bug in which we entered an infinite loop. The bug was not related to the button, it was already on main, because I was using wrongly Unity.Random(), which didn't include the second parameter in the range of randoms to return, causing to never find an available animation index, therefore entering into a loop when a team had 4 players.

In addition, the array of animations index was not setted to false in the index of the user switching team, making it enter in the infinite loop again when pressing the switch button 8 times.

All that is now fixed :)